% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchardet-package.R
\docType{package}
\name{uchardet}
\alias{uchardet}
\alias{uchardet-package}
\title{The Universal Character Encoding Detector}
\description{
R bindings for the uchardet library
(<https://www.freedesktop.org/wiki/Software/uchardet/>), that
is the encoding detector library of Mozilla. It takes a sequence
of bytes in an unknown character encoding without any additional
information, and attempts to determine the encoding of the text.
Returned encoding names are iconv-compatible.
}
\references{
\code{uchardet} page: \url{https://www.freedesktop.org/wiki/Software/uchardet/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://artemklevtsov.gitlab.io/uchardet}
  \item \url{https://gitlab.com/artemklevtsov}
  \item Report bugs at \url{https://gitlab.com/artemklevtsov/uchardet/issues}
}

}
\author{
\strong{Maintainer}: Artem Klevtsov \email{a.a.klevtsov@gmail.com} (\href{https://orcid.org/0000-0003-0492-6647}{ORCID})

Other contributors:
\itemize{
  \item Philipp Upravitelev \email{upravitelev@gmail.com} [contributor]
}

}
